[ALPS04239425] DTS: Peripheral porting
[Detail]
1.dtsi porting
2.kernelconfig refactor
3.driver refactor
  - accdet/sched
4.GPU driver
5.IMG sensor driver

MTK-Commit-Id: a7d71caf264ced24d58edbcb4b640c12d0e6a9de

CR-Id: ALPS04239425
Feature:[Module]Kernel Maintenance

+++ drivers/input/keyboard/mediatek/mt6739/hal_kpd.c
@@ -16,6 +16,7 @@
 #include <linux/of.h>
 #include <linux/of_address.h>
 #include <linux/of_irq.h>
+#include <mt-plat/mtk_boot_common.h>
 #ifdef CONFIG_MTK_PMIC_NEW_ARCH
 #include <mt-plat/upmu_common.h>
 #endif
@@ -310,7 +311,7 @@ void kpd_wakeup_src_setting(int enable)
 }
 
 /********************************************************************/
-void kpd_init_keymap(u16 keymap[])
+void kpd_init_keymap(u32 keymap[])
 {
 	int i = 0;
